requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.6
paddlepaddle==3.1.1
paddleocr==3.4.0
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==1.26.4
shapely==2.0.7
python-multipart==0.0.20
oss2==2.19.1
requests==2.32.3
PyYAML==6.0.2
